Fix test collection and ML dependency imports
- Move fixtures from tests/fixtures/conftest.py to tests/conftest.py
- Remove tests/fixtures directory (pytest plugin conflict)
- Add pytest.importorskip for ML-dependent test modules
- Remove ML processors from processing/__init__.py exports
- ML processors must be imported directly from submodules

This allows tests to run without ML dependencies installed.
ML-specific tests are skipped automatically.
